adding logging, moving add_memory to a memory python file to be imported
