+++ autogpts/ZEROAGPT_03/forge/sdk/abilities/google_search.py
@@ -6,7 +6,12 @@ from typing import List
 import os
 import googleapiclient.discovery
 
+from forge.sdk.memory.memstore_tools import add_memory
+
+from ..forge_log import ForgeLogger
 from .registry import ability
+
+logger = ForgeLogger(__name__)
 
 @ability(
     name="google_search",
@@ -25,21 +30,29 @@ async def google_search(agent, task_id: str, query: str) -> List[str]:
     """
     Return list of snippets from google search
     """
-    service = googleapiclient.discovery.build(
-        "customsearch",
-        "v1",
-        developerKey=os.getenv("GOOGLE_API_KEY"))
-    
-    response_snippets = []
-    response = service.cse().list(
-        q=query,
-        cx=os.getenv("GOOGLE_CSE_ID")
-    ).execute()
 
-    for result in response["items"]:
-        response_snippets.append({
-            "url": result["formattedUrl"],
-            "snippet": result["snippet"]
-        })
+    response_snippets = []
+
+    try:
+        service = googleapiclient.discovery.build(
+            "customsearch",
+            "v1",
+            developerKey=os.getenv("GOOGLE_API_KEY"))
+        
+        
+        response = service.cse().list(
+            q=query,
+            cx=os.getenv("GOOGLE_CSE_ID")
+        ).execute()
+
+        for result in response["items"]:
+            response_snippets.append({
+                "url": result["formattedUrl"],
+                "snippet": result["snippet"]
+            })
+
+        add_memory(task_id, response_snippets, "google_search")
+    except Exception as err:
+        logger.error(f"google_search failed: {err}")
 
     return response_snippets
