+++ autogpts/ZEROAGPT_03/forge/sdk/abilities/memories.py
@@ -4,9 +4,12 @@ Memory tool for document search
 from typing import List
 
 import os
+from ..forge_log import ForgeLogger
 from .registry import ability
 
 from forge.sdk.memory.memstore import ChromaMemStore
+
+logger = ForgeLogger(__name__)
 
 @ability(
     name="mem_search",
@@ -22,14 +25,21 @@ from forge.sdk.memory.memstore import ChromaMemStore
     output_type="list[str]",
 )
 async def mem_search(agent, task_id: str, query: str) -> List[str]:
-    chromadb_path = f"{os.getenv('AGENT_WORKSPACE')}/{task_id}"
+    mem_docs = []
 
-    memory = ChromaMemStore(chromadb_path)
-    memory_resp = memory.query(
-        task_id=task_id,
-        query=query
-    )
+    try:
+        chromadb_path = f"{os.getenv('AGENT_WORKSPACE')}/{task_id}"
 
-    return memory_resp["documents"][0]
+        memory = ChromaMemStore(chromadb_path)
+        memory_resp = memory.query(
+            task_id=task_id,
+            query=query
+        )
+
+        mem_docs = memory_resp["documents"][0]
+    except Exception as err:
+        logger.error(f"mem_search filed: {err}")
+
+    return mem_docs
